+++ songs/raleigh-and-spencer.js
@@ -1,0 +1,134 @@
+registerSong({
+    name: "raleigh-and-spencer",
+    title: "Raleigh & Spencer",
+    author: "Traditional",
+    tuning: "triplec",
+    categories: [ "banjo"],
+    tempo: 120,
+    chords: [ ],
+    tab: [
+      { 
+        name: "A",
+        signature: "2/4",
+        measures: [
+          {
+            groups: [
+              { notes: [ "Q" ] },
+              { notes: [ "S30", "S33H", "S33", "S30P" ] }
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E20", "S20", "S50" ] },
+              { notes: [ "S10g", "S20+T", "S20", "S50" ] }
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E23b", "S23g", "S50" ] },
+              { notes: [ "E20", "S10,S20,S30", "S50" ] }
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E20", "S20", "S50" ] },
+              { notes: [ "S20g", "S33+T", "S10,S20,S33", "S50" ] }
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E20", "S10,S20,S30", "S50" ] },
+              { notes: [ "S44", "S45H", "S20,S30,S45", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E30", "S10,S20,S30", "S50" ] },
+              { notes: [ "E33b", "S33g", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E30", "S10,S20,S30", "S50" ] },
+              { notes: [ "S43", "S40P", "E42" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E40", "S10,S20,S30", "S50" ] },
+              { notes: [ "S40", "S43H", "S10,S20,S30", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E40", "S40", "S50" ] },
+              { notes: [ "E44", "S45", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E30", "S10,S20,S30", "S50" ] },
+              { notes: [ "E33b", "S33g", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E30", "S10,S20,S30", "S50" ] },
+              { notes: [ "S43", "S40P", "E42" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E40", "S10,S20,S30", "S50" ] },
+              { notes: [ "S40g", "S40+T", "S20,S30,S40", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+              { notes: [ "E40", "S10,S20,S30", "S50" ] },
+              { notes: [ "S40g", "S40+T", "S20,S30,S40", "S50" ] },
+            ]
+          },
+
+          {
+            groups: [
+                { notes: [ "E112", "S110", "S50" ] },
+                { notes: [ "E112", "S110", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+                { notes: [ "E112", "S110", "S50" ] },
+                { notes: [ "S110g", "S29", "S110", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+                { notes: [ "E112", "S110", "S50" ] },
+                { notes: [ "E112", "S110", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+                { notes: [ "E112", "S110", "S50" ] },
+                { notes: [ "S110g", "S29", "S110", "S50" ] },
+            ]
+          },
+
+          {
+            groups: [
+                { notes: [ "E17", "S17g", "S50" ] },
+                { notes: [ "E110", "S110g", "S50" ] },
+            ]
+          },
+          {
+            groups: [
+                { notes: [ "E17", "S17", "S50" ] },
+                { notes: [ "S17", "S25+T", "S23", "S50" ] },
+            ]
+          },
+      ]
+      }
+    ],
+  });
+